requirements.txt
fastapi
uvicorn[standard]
python-dotenv
jinja2
google-genai
mistralai
requests
python-multipart